requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
selenium
webdriver-manager
openai-whisper>=20231117
pyaudio>=0.2.11
numpy>=1.24.0
torch>=2.0.0
torchaudio>=2.0.0

